gpc-bid-bot-minimal (1)/gpc-bid-bot/pyproject.toml
[project]
name = "gpc-bid-bot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["pydantic>=2.7","PyYAML>=6.0"]
[project.optional-dependencies]
dev = ["pytest>=8.2","pytest-cov>=5.0"]
[tool.pytest.ini_options]
addopts = "-q --cov=packages --cov-report=term-missing"
pythonpath = ["."]
